Add overrides to use plural types
There’s a feature in JaSerializer 0.13 to apply this globally
but when I tried to use it I was pulled into a cascading
problem with these issues:
https://github.com/phoenixframework/phoenix/issues/3119
https://github.com/getsentry/sentry-elixir/issues/275

I couldn’t figure that out so I’m manually overriding instead
for now.

This is all happening because Mirage uses plural types and
the overlaps-processing client-side code needed to check
for both, ugh.